requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
duckduckgo-search==6.2.13
langchain==0.2.16
langchain-core==0.2.40
langchain-community==0.2.16
langchain-groq==0.1.9
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
tenacity==8.5.0
